requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.5.3
billiard==3.3.0.23
cairocffi==0.8.0
CairoSVG==2.1.3
celery==3.1.18
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cssselect2==0.2.1
Cython==0.27.3
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0
django-bootstrap==0.2.4
django-bootstrap3==9.1.0
django-bootstrap4==0.0.5
django-db-log==2.2.1
django-db-logger==0.1.6
django-paypal==0.4.1
django-semanticui-forms==1.6.5
flower==0.9.2
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
kombu==3.0.37
mercadopago==0.3.4
numpy==1.14.0
pdfrw==0.4
Pillow==5.0.0
psycopg2==2.7.4
pycparser==2.18
Pyphen==0.9.4
pystan==2.17.1.0
pytz==2017.3
requests==2.18.4
six==1.11.0
South==1.0.2
static3==0.7.0
tinycss2==0.6.1
tornado==4.5.3
urllib3==1.22
WeasyPrint==0.42.2
webencodings==0.5.1
whitenoise==3.3.1
